+++ contracts/p0/RevenueTrader.sol
@@ -83,7 +83,7 @@ contract RevenueTraderP0 is TradingP0, IRevenueTrader {
         main.assetRegistry().refresh();
 
         IAsset assetToBuy = main.assetRegistry().toAsset(tokenToBuy);
-        (uint192 buyLow, uint192 buyHigh) = assetToBuy.lotPrice(); // {UoA/tok}
+        (uint192 buyLow, uint192 buyHigh) = assetToBuy.price(); // {UoA/tok}
         require(buyHigh > 0 && buyHigh < FIX_MAX, "buy asset price unknown");
 
         // For each ERC20: start auction of given kind
@@ -99,7 +99,7 @@ contract RevenueTraderP0 is TradingP0, IRevenueTrader {
             require(address(trades[erc20]) == address(0), "trade open");
             require(erc20.balanceOf(address(this)) > 0, "0 balance");
 
-            (uint192 sellLow, uint192 sellHigh) = assetToSell.lotPrice(); // {UoA/tok}
+            (uint192 sellLow, uint192 sellHigh) = assetToSell.price(); // {UoA/tok}
 
             TradingLibP0.TradeInfo memory trade = TradingLibP0.TradeInfo({
                 sell: assetToSell,